Fold trash mutations into DocumentsController
